Add files via upload
This is a project that I have created, when I was teaching Power BI to Business Analysis in my time at Xander Talent. 

This is an interesting mini project, which looks into a streaming company called Flix2C. 

The data has been created by Mockaroo, and their website is the following

https://www.mockaroo.com/

Have a look at the finished dashboard and play around with it yourself. 

-- Paul Mampilly